Move existing code from local/reader pkg to local/reader/volume pkg
