+++ internal/pkg/service/buffer/storage/level/local/reader/volume/config.go
@@ -1,4 +1,4 @@
-package reader
+package volume
 
 import (
 	"time"
